refactor(web): modularize LookerStudioReportCard by creating sub-components and a custom hook for report handling
