+++ apps/web/src/features/data-marts/reports/list/components/LookerStudioReportCard/LookerStudioReportCardActionRight.tsx
@@ -1,0 +1,17 @@
+import { cn } from '@owox/ui/lib/utils';
+import type { BaseCardComponentProps } from './types';
+
+export function LookerStudioReportCardActionRight({
+  children,
+  className,
+  ...props
+}: BaseCardComponentProps) {
+  return (
+    <div
+      className={cn('flex flex-shrink-0 items-center justify-center self-center p-4', className)}
+      {...props}
+    >
+      {children}
+    </div>
+  );
+}
