+++ apps/web/src/features/data-marts/reports/list/components/LookerStudioReportCard/LookerStudioReportCardDescription.tsx
@@ -1,0 +1,14 @@
+import { cn } from '@owox/ui/lib/utils';
+import type { BaseCardComponentProps } from './types';
+
+export function LookerStudioReportCardDescription({
+  children,
+  className,
+  ...props
+}: BaseCardComponentProps) {
+  return (
+    <div className={cn('text-muted-foreground text-sm', className)} {...props}>
+      {children}
+    </div>
+  );
+}
