+++ apps/web/src/features/data-marts/reports/list/components/LookerStudioReportCard/LookerStudioReportCardTitle.tsx
@@ -1,0 +1,14 @@
+import { cn } from '@owox/ui/lib/utils';
+import type { BaseCardComponentProps } from './types';
+
+export function LookerStudioReportCardTitle({
+  children,
+  className,
+  ...props
+}: BaseCardComponentProps) {
+  return (
+    <div className={cn('text-md font-medium', className)} {...props}>
+      {children}
+    </div>
+  );
+}
