+++ apps/web/src/features/data-marts/reports/list/components/LookerStudioReportCard/index.ts
@@ -1,6 +1,7 @@
 export { LookerStudioReportCard } from './LookerStudioReportCard';
-export { LookerStudioReportCardActionLeft } from './LookerStudioReportCard';
-export { LookerStudioReportCardContent } from './LookerStudioReportCard';
-export { LookerStudioReportCardTitle } from './LookerStudioReportCard';
-export { LookerStudioReportCardDescription } from './LookerStudioReportCard';
-export { LookerStudioReportCardActionRight } from './LookerStudioReportCard';
+export { LookerStudioReportCardActionLeft } from './LookerStudioReportCardActionLeft';
+export { LookerStudioReportCardContent } from './LookerStudioReportCardContent';
+export { LookerStudioReportCardTitle } from './LookerStudioReportCardTitle';
+export { LookerStudioReportCardDescription } from './LookerStudioReportCardDescription';
+export { LookerStudioReportCardActionRight } from './LookerStudioReportCardActionRight';
+export * from './types';
