+++ apps/web/src/features/data-marts/reports/list/components/LookerStudioReportCard/types/index.ts
@@ -1,0 +1,5 @@
+import type { ComponentPropsWithoutRef, ReactNode } from 'react';
+
+export interface BaseCardComponentProps extends ComponentPropsWithoutRef<'div'> {
+  children: ReactNode;
+}
